move read_auxiliary to common.py, enable --aux for table
